+++ Examples/Resources/SqlAG/3-CreateAvailabilityGroupDetailed.ps1
@@ -45,7 +45,8 @@ Configuration Example
 
     Import-DscResource -ModuleName SqlServerDsc
 
-    Node $AllNodes.NodeName {
+    Node $AllNodes.NodeName
+    {
         # Adding the required service account to allow the cluster to log into SQL
         SqlServerLogin AddNTServiceClusSvc
         {
@@ -80,6 +81,14 @@ Configuration Example
             PsDscRunAsCredential = $SqlAdministratorCredential
         }
 
+        SqlAlwaysOnService EnableHADR
+        {
+            Ensure               = 'Present'
+            InstanceName         = $Node.InstanceName
+            ServerName           = $Node.NodeName
+            PsDscRunAsCredential = $SqlAdministratorCredential
+        }
+
         if ( $Node.Role -eq 'PrimaryReplica' )
         {
             # Create the availability group on the instance tagged as the primary replica
@@ -104,7 +113,7 @@ Configuration Example
                 DatabaseHealthTrigger         = $Node.DatabaseHealthTrigger
                 DtcSupportEnabled             = $Node.DtcSupportEnabled
 
-                DependsOn                     = '[SqlServerEndpoint]HADREndpoint', '[SqlServerPermission]AddNTServiceClusSvcPermissions'
+                DependsOn                     = '[SqlAlwaysOnService]EnableHADR', '[SqlServerEndpoint]HADREndpoint', '[SqlServerPermission]AddNTServiceClusSvcPermissions'
                 PsDscRunAsCredential          = $SqlAdministratorCredential
             }
         }